fix: forzar float64 explícito en STORE_STOCK y PRICE para evitar Invalid value int64
Cuando el archivo Meli no tiene '-' en las columnas STORE_STOCK, pd.to_numeric
las deja como int64. Al asignar NaN del ERP (left join sin match) a int64
pandas lanza 'Invalid value [nan] for dtype int64'. El .astype(float) garantiza
float64 siempre, independiente del contenido del archivo.

+++ MarketMasterApp.py
@@ -466,11 +466,12 @@ def pagina_meli_av19_bulevar_cedi_oviedo():
                     merged_data['Original_Price'] = merged_data['PRICE']
                     merged_data['original_order'] = merged_data.index
 
-                    # Permitir NaNs en estas columnas convirtiéndolas a float temporalmente
-                    merged_data['PRICE'] = pd.to_numeric(merged_data['PRICE'], errors='coerce')
+                    # Permitir NaNs en estas columnas convirtiéndolas a float64 explícitamente
+                    # Forzar float64 (no solo to_numeric) para evitar int64 cuando no hay '-'
+                    merged_data['PRICE'] = pd.to_numeric(merged_data['PRICE'], errors='coerce').astype(float)
                     for col in merged_data.columns:
                         if "STORE_STOCK" in col:
-                            merged_data[col] = pd.to_numeric(merged_data[col], errors='coerce')
+                            merged_data[col] = pd.to_numeric(merged_data[col], errors='coerce').astype(float)
 
                     grouped = merged_data.groupby('ITEM_ID')
                     processed_groups = []
